Centralize creation of dockerflow endpoints.
