+++ auslib/admin/base.py
@@ -1,9 +1,5 @@
-from os import path
-
-from flask import Flask, request, jsonify, Response
+from flask import Flask, request
 from flask_compress import Compress
-
-import auslib
 
 import logging
 log = logging.getLogger(__name__)
@@ -19,7 +15,9 @@ from auslib.admin.views.releases import SingleLocaleView, \
 from auslib.admin.views.rules import RulesAPIView, \
     SingleRuleView, RuleHistoryAPIView
 from auslib.admin.views.history import DiffView, FieldView
-from auslib.global_state import dbo
+from auslib.dockerflow import create_dockerflow_endpoints
+
+create_dockerflow_endpoints(app)
 
 
 @app.errorhandler(500)
@@ -61,40 +59,3 @@ app.add_url_rule("/releases/<release>/builds/<platform>/<locale>", view_func=Sin
 app.add_url_rule("/releases/<release>/revisions", view_func=ReleaseHistoryView.as_view("release_revisions"))
 app.add_url_rule("/history/diff/<type_>/<change_id>/<field>", view_func=DiffView.as_view("diff"))
 app.add_url_rule("/history/view/<type_>/<change_id>/<field>", view_func=FieldView.as_view("field"))
-
-
-# Endpoints required by CloudOps as part of the Dockerflow spec: https://github.com/mozilla-services/Dockerflow
-
-
-@app.route("/__version__")
-def version():
-    version_file = app.config.get("VERSION_FILE")
-    if version_file and path.exists(version_file):
-        with open(app.config["VERSION_FILE"]) as f:
-            version_json = f.read()
-        return Response(version_json, mimetype="application/json")
-    else:
-        return jsonify({
-            "source": "https://github.com/mozilla/balrog",
-            "version": "unknown",
-            "commit": "unknown",
-        })
-
-
-@app.route("/__heartbeat__")
-def heartbeat():
-    """Per the Dockerflow spec:
-    Respond to /__heartbeat__ with a HTTP 200 or 5xx on error. This should
-    depend on services like the database to also ensure they are healthy."""
-    # Counting the rules should be a trivial enough operation that it won't
-    # cause notable load, but will verify that the database works.
-    dbo.rules.countRules()
-    return "OK!"
-
-
-@app.route("/__lbheartbeat__")
-def lbheartbeat():
-    """Per the Dockerflow spec:
-    Respond to /__lbheartbeat__ with an HTTP 200. This is for load balancer
-    checks and should not check any dependent services."""
-    return "OK!"
